+++ src/main/java/com/levelup/levelup_academy/Service/StatisticProService.java
@@ -1,7 +1,9 @@
 package com.levelup.levelup_academy.Service;
 
+import com.levelup.levelup_academy.Api.ApiException;
 import com.levelup.levelup_academy.DTO.StatisticProDTO;
 import com.levelup.levelup_academy.Model.Pro;
+import com.levelup.levelup_academy.Model.StatisticChild;
 import com.levelup.levelup_academy.Model.StatisticPlayer;
 import com.levelup.levelup_academy.Model.StatisticPro;
 import com.levelup.levelup_academy.Repository.ProRepository;
@@ -9,20 +11,30 @@ import com.levelup.levelup_academy.Repository.StatisticProRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.Comparator;
+import java.util.List;
+import java.util.stream.Collectors;
+
 @Service
 @RequiredArgsConstructor
 public class StatisticProService {
     private final StatisticProRepository statisticProRepository;
     private final ProRepository proRepository;
 
-    public StatisticPro getStatisticById(Integer statId) {
-        return statisticProRepository.findById(statId)
-                .orElseThrow(() -> new RuntimeException("Statistic not found"));
+    public StatisticPro getStatisticsByProfessionalId(Integer professionalId) {
+        StatisticPro stat = statisticProRepository.findByPro_Id(professionalId);
+        if (stat == null) throw new ApiException("Statistic not found for this professional");
+        return stat;
     }
+
+    public List<StatisticPro> getAllStatisticsByTrainerId(Integer trainerId) {
+        return statisticProRepository.findByPro_Trainer_Id(trainerId);
+    }
+
 
     public void createStatistic(Integer proId, StatisticProDTO dto) {
         Pro pro = proRepository.findById(proId)
-                .orElseThrow(() -> new RuntimeException("Pro not found"));
+                .orElseThrow(() -> new ApiException("Pro not found"));
 
         StatisticPro stat = new StatisticPro(null, dto.getRate(), dto.getWinGame(), dto.getLossGame(),
                 dto.getTrophy(), dto.getField(), dto.getDate(), pro);
@@ -31,8 +43,10 @@ public class StatisticProService {
     }
 
     public void updateStatistic(Integer statId, StatisticProDTO dto) {
-        StatisticPro stat = statisticProRepository.findById(statId)
-                .orElseThrow(() -> new RuntimeException("Statistic not found"));
+        StatisticPro stat = statisticProRepository.findStatisticProById(statId);
+        if (stat==null){
+            throw new ApiException("Statistic not found");
+        }
 
         stat.setRate(dto.getRate());
         stat.setWinGame(dto.getWinGame());
@@ -44,9 +58,36 @@ public class StatisticProService {
         statisticProRepository.save(stat);
     }
     public void deleteStatistic(Integer statId) {
-        StatisticPro stat = statisticProRepository.findById(statId)
-                .orElseThrow(() -> new RuntimeException("Statistic not found"));
+        StatisticPro stat = statisticProRepository.findStatisticProById(statId);
+        if (stat==null){
+            throw new ApiException("Statistic not found");
+        }
 
         statisticProRepository.delete(stat);
+    }
+    public StatisticPro getProWithTopTrophy() {
+        List<StatisticPro> all = statisticProRepository.findAll();
+
+        return all.stream()
+                .filter(pro -> pro.getTrophy() != null)
+                .max(Comparator.comparingInt(pro -> getTrophyRank(pro.getTrophy())))
+                .orElseThrow(() -> new ApiException("No child has a trophy"));
+    }
+
+    private int getTrophyRank(String trophy) {
+        return switch (trophy.toUpperCase()) {
+            case "GOLD" -> 3;
+            case "SILVER" -> 2;
+            case "BRONZE" -> 1;
+            default -> 0;
+        };
+    }
+
+    public List<StatisticPro> getTop5ProByGame(Integer winGame) {
+        List<StatisticPro> all = statisticProRepository.findStatisticProByWinGame(winGame);
+        return all.stream()
+                .sorted(Comparator.comparing(StatisticPro::getRate).reversed())
+                .limit(5)
+                .collect(Collectors.toList());
     }
 }
